+++ src/api/control/iterative/iterativePosPidController.cpp
@@ -7,7 +7,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-#include "okapi/impl/control/iterative/iterativePosPidController.hpp"
+#include "okapi/api/control/iterative/iterativePosPidController.hpp"
 #include <algorithm>
 #include <cmath>
